Purge old contactgroup format from application code
Looks like this also fixes a data inconsistency in the show host
endpoint of the REST API, which returned an invalid value <tt>{}</tt>
for the contactgroup field for the effective attributes of all hosts
which inherit the builtin contact group default value instead of
returning a correct value for the contact group field according to the
schema.

According to the schema the value needs to have at least the
groups key set, like this: <tt>{"groups": []}</tt>.
